Don't use the before shadow element to hack the gradient border
Because it'd cause random rendering issues in Google Chrome's
built-in PDF viewer, which would look like the issue is in the PDF
itself and would hide the content that should be bordered.
Now, there still are random rendering issues, but they look more like
GPU/OS/Chrome issues, which are more likely to prompt the user to
refresh the page rather than move away.